requirements.txt
numpy
torch
torchvision
scikit-learn
matplotlib
Pillow
pandas
tqdm
